Assignee in list detail
